% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{computeBits}
\alias{computeBits}
\title{Calcualte bits}
\usage{
computeBits(pwm, N = 4, Nseqs)
}
\arguments{
\item{pwm}{Position weight matrix}

\item{N}{Number of letters in namespace}

\item{Nseqs}{Number of sequences in PWM}
}
\description{
Calcualte bits
}

