% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{findNamespace}
\alias{findNamespace}
\title{Find namespace}
\usage{
findNamespace(letter_mat, seq_type, namespace)
}
\arguments{
\item{letter_mat}{Matrix of latters}

\item{seq_type}{Sequence type}

\item{namespace}{Alphabet}
}
\description{
Find namespace
}

