% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_schemes.r
\name{get_col_scheme}
\alias{get_col_scheme}
\title{Get color scheme}
\usage{
get_col_scheme(col_scheme, seq_type = "auto")
}
\arguments{
\item{col_scheme}{name of color scheme}

\item{seq_type}{sequence type of color scheme
Color scheme based on chemistry of amino acids
Hydrophobicity index (PMID: 7108955) from -4.5 to 4.5}
}
\description{
Get color scheme
}

