% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\name{get_font}
\alias{get_font}
\title{Read font from file if not in global envir.}
\usage{
get_font(font)
}
\arguments{
\item{font}{font id as index of all_fonts vector}
}
\description{
Read font from file if not in global envir.
}

