% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{getHeightData}
\alias{getHeightData}
\title{Gets height data used to plot logo, uses helper function computeHeights}
\usage{
getHeightData(pwm, method, decreasing = T)
}
\arguments{
\item{pwm}{position weight matrix}

\item{unit}{height method}
}
\description{
Gets height data used to plot logo, uses helper function computeHeights
}

