% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{guessSeqType}
\alias{guessSeqType}
\title{Guess sequence type based on letter matrix}
\usage{
guessSeqType(sp)
}
\arguments{
\item{sp}{letters}
}
\description{
Guess sequence type based on letter matrix
}

