% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{letter_G}
\alias{letter_G}
\title{Generate a data frame for a polygon of "G" character}
\usage{
letter_G(x.pos, y.pos, ht, wt, id = NULL)
}
\arguments{
\item{x.pos}{The x-coordinate of the bottom left of the polygon.}

\item{y.pos}{The y-coordinate of the bottom left of the polygon.}

\item{ht}{The height of the polygon.}

\item{wt}{The width of the polygon.}
}
\value{
The data frame for the polygon of "G" character.
}
\description{
Generate a data frame for a polygon of "G" character
}
\examples{
letter_G(0, 0, 1, 1)
}

