% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{letterMatrix}
\alias{letterMatrix}
\title{Generate letter matrix from vector of sequences}
\usage{
letterMatrix(input)
}
\arguments{
\item{input}{vector of sequences}
}
\description{
Generate letter matrix from vector of sequences
}

