% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_schemes.r
\name{list_col_schemes}
\alias{list_col_schemes}
\title{List color schemes available in ggseqlogo}
\usage{
list_col_schemes(v = T)
}
\arguments{
\item{v}{If true, font names are printed to stderr. Otherwise, color scheme names are returned as a character vector}
}
\description{
List color schemes available in ggseqlogo
}
