% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\name{list_fonts}
\alias{list_fonts}
\title{List fonts available in ggseqlogo}
\usage{
list_fonts(v = T)
}
\arguments{
\item{v}{If true, font names are printed to stderr. Otherwise, font names are returned as a character vector}
}
\description{
List fonts available in ggseqlogo
}
