% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\name{logo_data}
\alias{logo_data}
\title{Generate height data for logo}
\usage{
logo_data(seqs, method = "bits", stack_width = 0.95, rev_stack_order = F,
  font = 1, seq_group = 1, seq_type = "auto", namespace = NULL)
}
\description{
Generate height data for logo
}

