% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.r
\name{makePFM}
\alias{makePFM}
\title{Construct relative frequency matrix}
\usage{
makePFM(seqs, seq_type = "auto", priors = "eq", namespace = NULL)
}
\arguments{
\item{seqs}{aligned sequences as vector}

\item{seq_type}{sequence type}

\item{priors}{piror probabilities}

\item{namespace}{letters used for matrix construction}
}
\description{
Construct relative frequency matrix
}

