% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\name{newRange}
\alias{newRange}
\title{Change range of values}
\usage{
newRange(old_vals, new_min = 0, new_max = 1)
}
\arguments{
\item{old_vals}{values to transform}

\item{new_min}{minimum range of transformed data}

\item{new_max}{maximum range of transformed data}
}
\description{
Change range of values
}

