% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\docType{data}
\name{pfms_dna}
\alias{pfms_dna}
\title{List of position frequency matrices for transcription factors}
\description{
List of position frequency matrices for transcription factors
}
\keyword{data}
