% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqlogo.r
\name{theme_logo}
\alias{theme_logo}
\title{ggseqlogo custom theme}
\usage{
theme_logo(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{font base size}

\item{base_family}{font base family}
}
\description{
ggseqlogo custom theme
}
